requirements.txt
torch>=1.9.0
numpy>=1.19.0
scipy>=1.5.0
scikit-learn>=0.23.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.1.0
pyyaml>=5.3.0
tqdm>=4.50.0
jupyter>=1.0.0
pywt>=1.1.0  # For wavelet transforms